Remove write_snmp_service
