+++ cloudshell/snmp/cloudshell_snmp.py
@@ -19,14 +19,12 @@ class Snmp(object):
     def get_snmp_service(self, snmp_timeout=SNMP_TIMEOUT, snmp_retry_count=SNMP_RETRIES_COUNT):
         snmp_engine = self._get_snmp_engine(snmp_timeout, snmp_retry_count)
         snmp_context = SnmpContext(self._v3_context_engine_id, self._v3_context_name)
-        is_read_only = type(self._snmp_parameters) is SNMPV2ReadParameters
         get_bulk_flag = self._snmp_parameters.snmp_version > 0
         return SnmpContextManager(snmp_engine=snmp_engine,
                                   v3_context_engine_id=snmp_context.context_engine_id,
                                   v3_context_name=snmp_context.context_name,
                                   logger=self._logger,
-                                  get_bulk_flag=get_bulk_flag,
-                                  is_read_only=is_read_only
+                                  get_bulk_flag=get_bulk_flag
                                   )
 
     def _get_snmp_engine(self, snmp_timeout=SNMP_TIMEOUT, snmp_retry_count=SNMP_RETRIES_COUNT):
